Added ImageRenderer to standard richtext library
This commit adds a new embedded image renderer to the richtext library.
By default, images will only be output as a string (for example
"Asset#213zsd2"). When the renderer is enabled, it will instead output
an HTML image tag. Since this is possibly breaking existing dependent
code, this renderer must be manually enabled.