Added localization to richtext parsing
This commit adds correct localization to richtext parsing. Previously,
the parser wasn't given a locale. This lead to bugs when the parser
attempted to resolve embedded entries or assets which were localized.
Note that this changes the ParserInterface and the NodeMapperInterface,
which means that this is a backwards-incompatible change and might need
adjustments for custom types.

See issue #65